move visual pass detail computation to separate function
