+++ src/passpredict/observers/functions.py
@@ -5,6 +5,17 @@ import datetime
 import numpy as np
 
 from ..time import julian_date
+from .core import PassType
+from ..constants import R_EARTH
+
+if typing.TYPE_CHECKING:
+    from .observer import Observer
+
+
+class VisualPoints(typing.NamedTuple):
+    vis_begin_mjd: float = None
+    vis_tca_mjd: float = None
+    vis_end_mjd: float = None
 
 
 def make_utc(d: datetime.datetime) -> datetime.datetime:
@@ -78,3 +89,77 @@ def find_min(f: typing.Callable, a: float, b: float, tol: float) -> float:
 def julian_date_sum(d: datetime.datetime) -> float:
     jd, jdfr = julian_date(d.year, d.month, d.day, d.hour, d.minute, d.second + d.microsecond/1e6)
     return jd + jdfr
+
+
+def visual_pass_details(
+    observer: Observer,
+    aos_mjd: float,
+    tca_mjd: float,
+    los_mjd: float,
+    *,
+    tol: float = 1/86400.0,
+    sunrise_dg: float = -6,
+    n: int = 5,  # number of steps
+) -> typing.Tuple([PassType, VisualPoints]):
+    t0 = aos_mjd
+    tf = los_mjd
+    t = np.linspace(t0, tf, n)
+
+    def sun_el_fn(t):
+        return observer.location.sun_elevation_mjd(t) - sunrise_dg
+
+    el = np.array([sun_el_fn(t_) for t_ in t])
+
+    if np.min(el) > 0:
+        # entire pass in sunlit
+        return (PassType.daylight, VisualPoints(None, None, None))
+
+    if el[0]*el[-1] < 0:
+        # part of the pass is in darkness.
+        # only part of the pass is sunlit. Find new jd0, jdf
+        x = find_root(sun_el_fn, t0, tf, tol)
+        tmp1 = sun_el_fn(x - tol)
+        tmp2 = sun_el_fn(x + tol)
+        if tmp1 < tmp2:
+            # sun elevation is decreasing
+            t0 = x
+        else:
+            tf = x
+    # Now use t0 and tf to find when satellite is
+    # illuminated by sun
+    t = np.linspace(t0, tf, n)
+
+    def illum_fn(t):
+        return observer.satellite.illumination_distance_mjd(t) - R_EARTH  # noqa
+
+    illum_pts = np.array([illum_fn(t_) for t_ in t])
+
+    if np.max(illum_pts) < 0:
+        # entire pass is in shadow
+        return (PassType.unlit, VisualPoints(None, None, None))
+
+    if illum_pts[0]*illum_pts[-1] < 0:
+        # the satellite is visible for only part of the pass.
+        # Find new t0, tf
+        x = find_root(illum_fn, t0, tf, tol)
+        tmp1 = illum_fn(x - tol)
+        tmp2 = illum_fn(x + tol)
+        if tmp1 < tmp2:
+            # satellite is coming out of shadow
+            t0 = x
+        else:
+            # satellite is going into shadow
+            tf = x
+
+    # Set visible start and end points for Pass
+    vis_begin_mjd = t0
+    vis_end_mjd = tf
+    # Find maximum elevation during visible period
+    if vis_begin_mjd <= tca_mjd <= vis_end_mjd:
+        vis_tca_mjd = tca_mjd
+    elif observer._elevation_at_mjd(vis_begin_mjd) > observer._elevation_at_mjd(vis_end_mjd):
+        vis_tca_mjd = vis_begin_mjd
+    else:
+        vis_tca_mjd = vis_end_mjd
+    visual_points = VisualPoints(vis_begin_mjd, vis_tca_mjd, vis_end_mjd)
+    return (PassType.visible, visual_points)
